+++ ui/src/app/core/manage-plugins/plugin-logs/plugin-logs.component.ts
@@ -69,7 +69,7 @@ export class PluginLogsComponent implements OnInit, OnDestroy {
     ref.componentInstance.title = this.$translate.instant('logs.title_download_log_file')
     ref.componentInstance.message = this.$translate.instant('logs.download_warning')
     ref.componentInstance.confirmButtonLabel = this.$translate.instant('form.button_download')
-    ref.componentInstance.faIconClass = 'fas fa-fw fa-user-secret primary-text'
+    ref.componentInstance.faIconClass = 'fas fa-user-secret primary-text'
 
     ref.result
       .then(() => {