+++ ui/src/app/modules/plugins/plugin-card/plugin-info/plugin-info.component.ts
@@ -17,8 +17,8 @@ export class PluginInfoComponent implements OnInit {
   @Input() plugin: Plugin
 
   public readonly defaultIcon = 'assets/hb-icon.png'
-  public readonly linkScoped = '<a href="https://github.com/homebridge/plugins/wiki/Scoped-Plugins" target="_blank"><i class="fas fa-fw fa-external-link-alt primary-text"></i></a>'
-  public readonly linkVerified = '<a href="https://github.com/homebridge/plugins/wiki/Verified-Plugins" target="_blank"><i class="fas fa-fw fa-external-link-alt primary-text"></i></a>'
+  public readonly linkScoped = '<a href="https://github.com/homebridge/plugins/wiki/Scoped-Plugins" target="_blank"><i class="fas fa-external-link-alt primary-text"></i></a>'
+  public readonly linkVerified = '<a href="https://github.com/homebridge/plugins/wiki/Verified-Plugins" target="_blank"><i class="fas fa-external-link-alt primary-text"></i></a>'
 
   public ngOnInit() {
     if (!this.plugin.icon) {